Fixes for the build
- The logic for editing water in the editor has been moved to a separate file and is NOT included in the build
- Unsuccessful developments in the pixelation shader have been removed
