+++ Assets/Scripts/Water/Editor.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 4b12261d4fbebed4aa8c5e0cc07b6182
-NativeFormatImporter:
+guid: d889b791d5f29f7488c473297863a960
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
