+++ Assets/Scripts/Water/Editor/InteractableWaterEditor.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 83d5e2d3d33edea41820da79905afdb5
+guid: 8ecb321b04c2c2944ad7b58ea05d0ef4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
