--- Assets/Shaders/Pixelate/Pixelate.shader.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: dc7b48d8dbf00f34ba2e5d2756d5d6e0
-ShaderImporter:
-  externalObjects: {}
-  defaultTextures: []
-  nonModifiableTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
